Change int use to size_t inside ntl.h and ntl.c (#288)
* Change int use to size_t inside ntl.h and ntl.c

* Change int to size_t inside comment

+++ ntl.h
@@ -32,7 +32,7 @@
  * p[n] points to NULL
  *
  *
- *  for (int i = 0; p[i]; i++)
+ *  for (size_t i = 0; p[i]; i++)
  *    // do something here for each element
  *
  */
@@ -134,12 +134,12 @@ struct ntl_str_delimiter {
  * ntl_to_buf(buf, n, ..) serialize p to buf and return the number of
  *       bytes written excluding \0
  */
-int ntl_to_buf(char *buf, size_t buf_size, void **p,
+size_t ntl_to_buf(char *buf, size_t buf_size, void **p,
                struct ntl_str_delimiter  * d,
                ntl_elem_serializer * x);
 
 
-int ntl_to_abuf(char **buf_ptr, void **p, struct ntl_str_delimiter  * d,
+size_t ntl_to_abuf(char **buf_ptr, void **p, struct ntl_str_delimiter  * d,
                 ntl_elem_serializer * x);
